Revert "Move kbucket into separate DHT crate"
This reverts commit d47726daca8090cccd4856b54d107b63ec952c4f.
