--- discv5-dht/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "discv5-dht"
-authors = ["Age Manning <Age@AgeManning.com>"]
-edition = "2018"
-version = "0.1.0"
-description = "Implementation of Kademlia DHT"
-license = "MIT"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-arrayvec = "0.7.1"
-enr = { version = "0.5.1", features = ["k256", "ed25519"] }
-sha2 = "0.9.5"
-tracing = { version = "0.1.26", features = ["log"] }
-uint = { version = "0.9.1", default-features = false }
-
-[dev-dependencies]
-rand_07 = { package = "rand", version = "0.7" }
-quickcheck = "0.9.2"
