--- discv5-dht/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod kbucket;
-pub type Enr = enr::Enr<enr::CombinedKey>;
-// re-export the ENR crate
-pub use enr;
