--- discv5/Cargo.toml
@@ -1,54 +1,0 @@
-[package]
-name = "discv5"
-authors = ["Age Manning <Age@AgeManning.com>"]
-edition = "2018"
-version = "0.1.0-beta.9"
-description = "Implementation of the p2p discv5 discovery protocol"
-license = "MIT"
-repository = "https://github.com/sigp/discv5"
-readme = "../README.md"
-keywords = ["peer-to-peer", "libp2p", "networking", "discovery", "discv5"]
-categories = ["network-programming", "asynchronous"]
-exclude = [".gitignore", ".github/*"]
-
-[dependencies]
-discv5-dht = { path = "../discv5-dht" }
-enr = { version = "0.5.1", features = ["k256", "ed25519"] }
-tokio = { version = "1.7.1", features = ["net", "sync", "macros", "rt"] }
-tokio-stream = "0.1.6"
-tokio-util = { version = "0.6.7", features = ["time"] }
-libp2p-core = { version = "0.29.0", optional = true }
-zeroize = { version = "1.3.0", features = ["zeroize_derive"] }
-futures = "0.3.15"
-uint = { version = "0.9.1", default-features = false }
-rlp = "0.5.0"
-sha2 = "0.9.5"
-hkdf = "0.11.0"
-hex = "0.4.3"
-fnv = "1.0.7"
-arrayvec = "0.7.1"
-digest = "0.9.0"
-rand = { version = "0.8.4", package = "rand" }
-smallvec = "1.6.1"
-parking_lot = "0.11.1"
-lru_time_cache = "0.11.11"
-lazy_static = "1.4.0"
-aes = { version = "0.7.4", features = ["ctr"] }
-aes-gcm = "0.9.2"
-tracing = { version = "0.1.26", features = ["log"] }
-tracing-subscriber = "0.2.19"
-lru = "0.6.5"
-
-[dev-dependencies]
-rand_07 = { package = "rand", version = "0.7" }
-quickcheck = "0.9.2"
-env_logger = "0.8.4"
-hex-literal = "0.3.1"
-simple_logger = "1.11.0"
-tokio-util = { version = "0.6.7", features = ["time"] }
-tokio = { version = "1.7.1", features = ["full"] }
-rand_xorshift = "0.3.0"
-rand_core = "0.6.3"
-
-[features]
-libp2p = ["libp2p-core"]
