+++ src/lib.rs
@@ -107,6 +107,7 @@ mod discv5;
 mod error;
 mod executor;
 pub mod handler;
+mod kbucket;
 pub mod metrics;
 mod node_info;
 pub mod packet;
@@ -125,6 +126,7 @@ pub use crate::discv5::{Discv5, Discv5Event};
 pub use config::{Discv5Config, Discv5ConfigBuilder};
 pub use error::{Discv5Error, QueryError, RequestError, ResponseError};
 pub use executor::{Executor, TokioExecutor};
+pub use kbucket::{ConnectionDirection, ConnectionState, Key};
 pub use permit_ban::PermitBanList;
 pub use service::TalkRequest;
 pub use socket::{RateLimiter, RateLimiterBuilder};
