+++ src/query_pool/peers/predicate.rs
@@ -1,36 +1,14 @@
-// Copyright 2018 Parity Technologies (UK) Ltd.
-//
-// Permission is hereby granted, free of charge, to any person obtaining a
-// copy of this software and associated documentation files (the "Software"),
-// to deal in the Software without restriction, including without limitation
-// the rights to use, copy, modify, merge, publish, distribute, sublicense,
-// and/or sell copies of the Software, and to permit persons to whom the
-// Software is furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-// DEALINGS IN THE SOFTWARE.
-
-// This basis of this file has been taken from the rust-libp2p codebase:
-// https://github.com/libp2p/rust-libp2p
-//
 use super::*;
-use crate::config::Discv5Config;
-use discv5_dht::kbucket::{Distance, Key, MAX_NODES_PER_BUCKET};
+use crate::{
+    config::Discv5Config,
+    kbucket::{Distance, Key, PredicateKey, MAX_NODES_PER_BUCKET},
+};
 use std::{
     collections::btree_map::{BTreeMap, Entry},
     time::{Duration, Instant},
 };
 
-#[derive(Debug, Clone)]
-pub struct FindNodeQuery<TNodeId> {
+pub(crate) struct PredicateQuery<TNodeId, TResult> {
     /// The target key we are looking for
     target_key: Key<TNodeId>,
 
@@ -43,26 +21,29 @@ pub struct FindNodeQuery<TNodeId> {
     /// The number of peers for which the query is currently waiting for results.
     num_waiting: usize,
 
+    /// The predicate function to be applied to filter the ENR's found during the search.
+    predicate: Box<dyn Fn(&TResult) -> bool + Send + 'static>,
+
     /// The configuration of the query.
-    config: FindNodeQueryConfig,
+    config: PredicateQueryConfig,
 }
 
 /// Configuration for a `Query`.
 #[derive(Debug, Clone)]
-pub struct FindNodeQueryConfig {
+pub(crate) struct PredicateQueryConfig {
     /// Allowed level of parallelism.
     ///
     /// The `α` parameter in the Kademlia paper. The maximum number of peers that a query
     /// is allowed to wait for in parallel while iterating towards the closest
     /// nodes to a target. Defaults to `3`.
-    pub parallelism: usize,
+    pub(crate) parallelism: usize,
 
     /// Number of results to produce.
     ///
     /// The number of closest peers that a query must obtain successful results
     /// for before it terminates. Defaults to the maximum number of entries in a
     /// single k-bucket, i.e. the `k` parameter in the Kademlia paper.
-    pub num_results: usize,
+    pub(crate) num_results: usize,
 
     /// The timeout for a single peer.
     ///
@@ -70,11 +51,11 @@ pub struct FindNodeQueryConfig {
     /// window, the iterator considers the peer unresponsive and will not wait for
     /// the peer when evaluating the termination conditions, until and unless a
     /// result is delivered. Defaults to `10` seconds.
-    pub peer_timeout: Duration,
+    pub(crate) peer_timeout: Duration,
 }
 
-impl FindNodeQueryConfig {
-    pub fn new_from_config(config: &Discv5Config) -> Self {
+impl PredicateQueryConfig {
+    pub(crate) fn new_from_config(config: &Discv5Config) -> Self {
         Self {
             parallelism: config.query_parallelism,
             num_results: MAX_NODES_PER_BUCKET,
@@ -83,27 +64,31 @@ impl FindNodeQueryConfig {
     }
 }
 
-impl<TNodeId> FindNodeQuery<TNodeId>
+impl<TNodeId, TResult> PredicateQuery<TNodeId, TResult>
 where
     TNodeId: Into<Key<TNodeId>> + Eq + Clone,
+    TResult: Into<TNodeId> + Clone,
 {
     /// Creates a new query with the given configuration.
     pub fn with_config<I>(
-        config: FindNodeQueryConfig,
+        config: PredicateQueryConfig,
         target_key: Key<TNodeId>,
         known_closest_peers: I,
+        predicate: impl Fn(&TResult) -> bool + Send + 'static,
     ) -> Self
     where
-        I: IntoIterator<Item = Key<TNodeId>>,
+        I: IntoIterator<Item = PredicateKey<TNodeId>>,
     {
         // Initialise the closest peers to begin the query with.
         let closest_peers = known_closest_peers
             .into_iter()
             .map(|key| {
-                let key: Key<TNodeId> = key;
+                let predicate_match = key.predicate_match;
+                let key: Key<TNodeId> = key.into();
                 let distance = key.distance(&target_key);
                 let state = QueryPeerState::NotContacted;
-                (distance, QueryPeer::new(key, state))
+
+                (distance, QueryPeer::new(key, state, predicate_match))
             })
             .take(config.num_results)
             .collect();
@@ -111,12 +96,13 @@ where
         // The query initially makes progress by iterating towards the target.
         let progress = QueryProgress::Iterating { no_progress: 0 };
 
-        FindNodeQuery {
+        PredicateQuery {
             config,
             target_key,
             progress,
             closest_peers,
             num_waiting: 0,
+            predicate: Box::new(predicate),
         }
     }
 
@@ -135,7 +121,10 @@ where
     /// If the query is finished, the query is not currently waiting for a
     /// result from `peer`, or a result for `peer` has already been reported,
     /// calling this function has no effect.
-    pub fn on_success(&mut self, node_id: &TNodeId, closer_peers: Vec<TNodeId>) {
+    pub fn on_success<'a>(&mut self, node_id: &TNodeId, closer_peers: &'a [TResult])
+    where
+        &'a TResult: Into<TNodeId>,
+    {
         if let QueryProgress::Finished = self.progress {
             return;
         }
@@ -172,10 +161,13 @@ where
         let num_closest = self.closest_peers.len();
 
         // Incorporate the reported closer peers into the query.
-        for peer in closer_peers {
-            let key: Key<TNodeId> = peer.into();
+        for result in closer_peers {
+            // If ENR satisfies the predicate, add to list of peers that satisfies predicate
+            let predicate_match = (self.predicate)(result);
+            let key: TNodeId = result.into();
+            let key: Key<TNodeId> = key.into();
             let distance = self.target_key.distance(&key);
-            let peer = QueryPeer::new(key, QueryPeerState::NotContacted);
+            let peer = QueryPeer::new(key, QueryPeerState::NotContacted, predicate_match);
             self.closest_peers.entry(distance).or_insert(peer);
             // The query makes progress if the new peer is either closer to the target
             // than any peer seen so far (i.e. is the first entry), or the query did
@@ -222,17 +214,12 @@ where
         let key: Key<TNodeId> = peer.clone().into();
         let distance = key.distance(&self.target_key);
 
-        match self.closest_peers.entry(distance) {
-            Entry::Vacant(_) => {}
-            Entry::Occupied(mut e) => match e.get().state {
-                QueryPeerState::Waiting(..) => {
-                    debug_assert!(self.num_waiting > 0);
-                    self.num_waiting -= 1;
-                    e.get_mut().state = QueryPeerState::Failed
-                }
-                QueryPeerState::Unresponsive => e.get_mut().state = QueryPeerState::Failed,
-                _ => {}
-            },
+        if let Entry::Occupied(mut e) = self.closest_peers.entry(distance) {
+            if let QueryPeerState::Waiting(..) = e.get().state {
+                debug_assert!(self.num_waiting > 0);
+                self.num_waiting -= 1;
+                e.get_mut().state = QueryPeerState::Failed
+            }
         }
     }
 
@@ -262,8 +249,8 @@ where
                         let timeout = now + self.config.peer_timeout;
                         peer.state = QueryPeerState::Waiting(timeout);
                         self.num_waiting += 1;
-                        let peer = peer.key.preimage().clone();
-                        return QueryState::Waiting(Some(peer));
+                        let return_peer = peer.key.preimage().clone();
+                        return QueryState::Waiting(Some(return_peer));
                     } else {
                         return QueryState::WaitingAtCapacity;
                     }
@@ -284,18 +271,23 @@ where
                         // `result_counter` did not yet reach `num_results`. Therefore
                         // the query is not yet done, regardless of already successful
                         // queries to peers farther from the target.
-                        result_counter = None;
+                        // Only count predicate peers.
+                        if peer.predicate_match {
+                            result_counter = None;
+                        }
                     }
                 }
 
                 QueryPeerState::Succeeded => {
                     if let Some(ref mut cnt) = result_counter {
-                        *cnt += 1;
-                        // If `num_results` successful results have been delivered for the
-                        // closest peers, the query is done.
-                        if *cnt >= self.config.num_results {
-                            self.progress = QueryProgress::Finished;
-                            return QueryState::Finished;
+                        if peer.predicate_match {
+                            *cnt += 1;
+                            // If `num_results` successful results have been delivered for the
+                            // closest peers, the query is done.
+                            if *cnt >= self.config.num_results {
+                                self.progress = QueryProgress::Finished;
+                                return QueryState::Finished;
+                            }
                         }
                     }
                 }
@@ -319,13 +311,17 @@ where
         }
     }
 
-    /// Consumes the query, returning the target and the closest peers.
+    /// Consumes the query, returning the peers who match the predicate.
     pub fn into_result(self) -> Vec<TNodeId> {
         self.closest_peers
             .into_iter()
             .filter_map(|(_, peer)| {
                 if let QueryPeerState::Succeeded = peer.state {
-                    Some(peer.key.into_preimage())
+                    if peer.predicate_match {
+                        Some(peer.key.into_preimage())
+                    } else {
+                        None
+                    }
                 } else {
                     None
                 }
@@ -392,22 +388,22 @@ struct QueryPeer<TNodeId> {
     /// The `KBucket` key used to identify the peer.
     key: Key<TNodeId>,
 
-    /// The current rpc request iteration that has been made on this peer.
-    iteration: usize,
-
     /// The number of peers that have been returned by this peer.
     peers_returned: usize,
+
+    /// Whether the peer has matched the predicate or not.
+    predicate_match: bool,
 
     /// The current query state of this peer.
     state: QueryPeerState,
 }
 
 impl<TNodeId> QueryPeer<TNodeId> {
-    pub fn new(key: Key<TNodeId>, state: QueryPeerState) -> Self {
+    pub fn new(key: Key<TNodeId>, state: QueryPeerState, predicate_match: bool) -> Self {
         QueryPeer {
             key,
-            iteration: 1,
             peers_returned: 0,
+            predicate_match,
             state,
         }
     }
@@ -439,271 +435,4 @@ enum QueryPeerState {
     ///
     /// This is a final state, reached as a result of a call to `on_success`.
     Succeeded,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use enr::NodeId;
-    use quickcheck::*;
-    use rand_07::{thread_rng, Rng};
-    use std::time::Duration;
-
-    type TestQuery = FindNodeQuery<NodeId>;
-
-    fn random_nodes(n: usize) -> impl Iterator<Item = NodeId> + Clone {
-        (0..n).map(|_| NodeId::random())
-    }
-
-    fn random_query<G: Rng>(g: &mut G) -> TestQuery {
-        let known_closest_peers = random_nodes(g.gen_range(1, 60)).map(Key::from);
-        let target = NodeId::random();
-        let config = FindNodeQueryConfig {
-            parallelism: g.gen_range(1, 10),
-            num_results: g.gen_range(1, 25),
-            peer_timeout: Duration::from_secs(g.gen_range(10, 30)),
-        };
-        FindNodeQuery::with_config(config, target.into(), known_closest_peers)
-    }
-
-    fn sorted(target: &Key<NodeId>, peers: &[Key<NodeId>]) -> bool {
-        peers
-            .windows(2)
-            .all(|w| w[0].distance(&target) < w[1].distance(&target))
-    }
-
-    impl Arbitrary for TestQuery {
-        fn arbitrary<G: Gen>(g: &mut G) -> TestQuery {
-            random_query(g)
-        }
-    }
-
-    #[test]
-    fn new_query() {
-        let query = random_query(&mut thread_rng());
-        let target = query.target_key.clone();
-
-        let (keys, states): (Vec<_>, Vec<_>) = query
-            .closest_peers
-            .values()
-            .map(|e| (e.key.clone(), &e.state))
-            .unzip();
-
-        let none_contacted = states
-            .iter()
-            .all(|s| matches!(s, QueryPeerState::NotContacted));
-
-        assert!(none_contacted, "Unexpected peer state in new query.");
-        assert!(
-            sorted(&target, &keys),
-            "Closest peers in new query not sorted by distance to target."
-        );
-        assert_eq!(
-            query.num_waiting, 0,
-            "Unexpected peers in progress in new query."
-        );
-        assert!(
-            query.into_result().is_empty(),
-            "Unexpected closest peers in new query"
-        );
-    }
-
-    #[test]
-    fn termination_and_parallelism() {
-        fn prop(mut query: TestQuery) {
-            let now = Instant::now();
-            let mut rng = thread_rng();
-
-            let mut expected = query
-                .closest_peers
-                .values()
-                .map(|e| e.key.clone())
-                .collect::<Vec<_>>();
-            let num_known = expected.len();
-            let max_parallelism = usize::min(query.config.parallelism, num_known);
-
-            let target = query.target_key.clone();
-            let mut remaining;
-            let mut num_failures = 0;
-
-            'finished: loop {
-                if expected.is_empty() {
-                    break;
-                }
-                // Split off the next up to `parallelism` expected peers.
-                else if expected.len() < max_parallelism {
-                    remaining = Vec::new();
-                } else {
-                    remaining = expected.split_off(max_parallelism);
-                }
-
-                // Advance the query for maximum parallelism.
-                for k in expected.iter() {
-                    match query.next(now) {
-                        QueryState::Finished => break 'finished,
-                        QueryState::Waiting(Some(p)) => assert_eq!(&p, k.preimage()),
-                        QueryState::Waiting(None) => panic!("Expected another peer."),
-                        QueryState::WaitingAtCapacity => panic!("Unexpectedly reached capacity."),
-                    }
-                }
-                let num_waiting = query.num_waiting;
-                assert_eq!(num_waiting, expected.len());
-
-                // Check the bounded parallelism.
-                if query.at_capacity() {
-                    assert_eq!(query.next(now), QueryState::WaitingAtCapacity)
-                }
-
-                // Report results back to the query with a random number of "closer"
-                // peers or an error, thus finishing the "in-flight requests".
-                for (i, k) in expected.iter().enumerate() {
-                    if rng.gen_bool(0.75) {
-                        let num_closer = rng.gen_range(0, query.config.num_results + 1);
-                        let closer_peers = random_nodes(num_closer).collect::<Vec<_>>();
-                        // let _: () = remaining;
-                        remaining.extend(closer_peers.iter().map(|x| Key::from(*x)));
-                        query.on_success(k.preimage(), closer_peers);
-                    } else {
-                        num_failures += 1;
-                        query.on_failure(k.preimage());
-                    }
-                    assert_eq!(query.num_waiting, num_waiting - (i + 1));
-                }
-
-                // Re-sort the remaining expected peers for the next "round".
-                remaining.sort_by_key(|k| target.distance(&k));
-
-                expected = remaining
-            }
-
-            // The query must be finished.
-            assert_eq!(query.next(now), QueryState::Finished);
-            assert_eq!(query.progress, QueryProgress::Finished);
-
-            // Determine if all peers have been contacted by the query. This _must_ be
-            // the case if the query finished with fewer than the requested number
-            // of results.
-            let all_contacted = query.closest_peers.values().all(|e| {
-                !matches!(
-                    e.state,
-                    QueryPeerState::NotContacted | QueryPeerState::Waiting { .. }
-                )
-            });
-
-            let target_key = query.target_key.clone();
-            let num_results = query.config.num_results;
-            let result = query.into_result();
-            let closest = result.into_iter().map(Key::from).collect::<Vec<_>>();
-
-            // assert_eq!(result.target, target);
-            assert!(sorted(&target_key, &closest));
-
-            if closest.len() < num_results {
-                // The query returned fewer results than requested. Therefore
-                // either the initial number of known peers must have been
-                // less than the desired number of results, or there must
-                // have been failures.
-                assert!(num_known < num_results || num_failures > 0);
-                // All peers must have been contacted.
-                assert!(all_contacted, "Not all peers have been contacted.");
-            } else {
-                assert_eq!(num_results, closest.len(), "Too  many results.");
-            }
-        }
-
-        QuickCheck::new().tests(10).quickcheck(prop as fn(_) -> _)
-    }
-
-    #[test]
-    fn no_duplicates() {
-        fn prop(mut query: TestQuery) -> bool {
-            let now = Instant::now();
-            let closer: Vec<NodeId> = random_nodes(1).collect();
-
-            // A first peer reports a "closer" peer.
-            let peer1 = if let QueryState::Waiting(Some(p)) = query.next(now) {
-                p
-            } else {
-                panic!("No peer.");
-            };
-            query.on_success(&peer1, closer.clone());
-            // Duplicate result from the same peer.
-            query.on_success(&peer1, closer.clone());
-
-            // If there is a second peer, let it also report the same "closer" peer.
-            match query.next(now) {
-                QueryState::Waiting(Some(p)) => {
-                    let peer2 = p;
-                    query.on_success(&peer2, closer.clone())
-                }
-                QueryState::Finished => {}
-                _ => panic!("Unexpectedly query state."),
-            };
-
-            // The "closer" peer must only be in the query once.
-            let n = query
-                .closest_peers
-                .values()
-                .filter(|e| e.key.preimage() == &closer[0])
-                .count();
-            assert_eq!(n, 1);
-
-            true
-        }
-
-        QuickCheck::new().tests(10).quickcheck(prop as fn(_) -> _)
-    }
-
-    #[test]
-    fn timeout() {
-        fn prop(mut query: TestQuery) -> bool {
-            let mut now = Instant::now();
-            let peer = query
-                .closest_peers
-                .values()
-                .next()
-                .unwrap()
-                .key
-                .clone()
-                .into_preimage();
-            // Poll the query for the first peer to be in progress.
-            match query.next(now) {
-                QueryState::Waiting(Some(id)) => assert_eq!(id, peer),
-                _ => panic!(),
-            }
-
-            // Artificially advance the clock.
-            now += query.config.peer_timeout;
-
-            // Advancing the query again should mark the first peer as unresponsive.
-            let _ = query.next(now);
-            match &query.closest_peers.values().next().unwrap() {
-                QueryPeer {
-                    key,
-                    state: QueryPeerState::Unresponsive,
-                    ..
-                } => {
-                    assert_eq!(key.preimage(), &peer);
-                }
-                QueryPeer { state, .. } => panic!("Unexpected peer state: {:?}", state),
-            }
-
-            let finished = query.progress == QueryProgress::Finished;
-            query.on_success(&peer, Vec::<NodeId>::new());
-            let closest = query.into_result();
-
-            if finished {
-                // Delivering results when the query already finished must have
-                // no effect.
-                assert_eq!(Vec::<NodeId>::new(), closest);
-            } else {
-                // Unresponsive peers can still deliver results while the iterator
-                // is not finished.
-                assert_eq!(vec![peer], closest)
-            }
-            true
-        }
-
-        QuickCheck::new().tests(10).quickcheck(prop as fn(_) -> _)
-    }
 }
