+++ src/service/query_info.rs
@@ -1,5 +1,4 @@
-use crate::{rpc::RequestBody, Enr};
-use discv5_dht::kbucket::Key;
+use crate::{kbucket::Key, rpc::RequestBody, Enr};
 use enr::NodeId;
 use sha2::digest::generic_array::GenericArray;
 use smallvec::SmallVec;
